requirement.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-environ==0.5.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
idna==3.2
install==1.3.4
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
openapi-codec==1.3.2
pkg-resources==0.0.0
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
simplejson==3.17.5
sqlparse==0.4.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
